Move SpellHelpers to API
