+++ src/main/java/vazkii/psi/common/spell/operator/entity/PieceOperatorClosestToLine.java
@@ -14,13 +14,13 @@ import net.minecraft.util.math.vector.Vector3d;
 import vazkii.psi.api.internal.Vector3;
 import vazkii.psi.api.spell.Spell;
 import vazkii.psi.api.spell.SpellContext;
+import vazkii.psi.api.spell.SpellHelpers;
 import vazkii.psi.api.spell.SpellParam;
 import vazkii.psi.api.spell.SpellRuntimeException;
 import vazkii.psi.api.spell.param.ParamEntityListWrapper;
 import vazkii.psi.api.spell.param.ParamVector;
 import vazkii.psi.api.spell.piece.PieceOperator;
 import vazkii.psi.api.spell.wrapper.EntityListWrapper;
-import vazkii.psi.common.core.helpers.SpellHelpers;
 
 import static vazkii.psi.common.spell.operator.entity.PieceOperatorClosestToPoint.closestToPoint;
