+++ src/main/java/vazkii/psi/common/spell/operator/entity/PieceOperatorEntityRaycast.java
@@ -17,12 +17,12 @@ import vazkii.psi.api.internal.Vector3;
 import vazkii.psi.api.spell.ISpellImmune;
 import vazkii.psi.api.spell.Spell;
 import vazkii.psi.api.spell.SpellContext;
+import vazkii.psi.api.spell.SpellHelpers;
 import vazkii.psi.api.spell.SpellParam;
 import vazkii.psi.api.spell.SpellRuntimeException;
 import vazkii.psi.api.spell.param.ParamNumber;
 import vazkii.psi.api.spell.param.ParamVector;
 import vazkii.psi.api.spell.piece.PieceOperator;
-import vazkii.psi.common.core.helpers.SpellHelpers;
 
 import java.util.Optional;
 import java.util.function.Predicate;
